#1328 Add editors for Grid_AddNewShapePoint and Grid_DrawShape

+++ EditorControls/NumberDoubleControl.xaml.cs
@@ -93,7 +93,7 @@ namespace TextAdventures.Quest.EditorControls
         public string StringValue
         {
             get { return ctlNumber.Value.Value.ToString(); }
-            set { ctlNumber.Value = int.Parse(value); }
+            set { ctlNumber.Value = double.Parse(value); }
         }
 
         public Control FocusableControl
